+++ ms-productos/src/main/java/isi/dan/ms_productos/exception/CategoriaNotFoundException.java
@@ -1,7 +1,7 @@
 package isi.dan.ms_productos.exception;
 
 public class CategoriaNotFoundException extends Exception{
-    public CategoriaNotFoundException(Long id){
+    public CategoriaNotFoundException(Integer id){
         super("Categoria "+id+" no encontrada");
     }
 }
